Add payment method summary to confirmation dialog
