+++ client/me/purchases/credit-cards/payment-method-delete-dialog/index.tsx
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import React, { FunctionComponent } from 'react';
-import { useTranslate } from 'i18n-calypso';
+import { useTranslate, TranslateResult } from 'i18n-calypso';
 
 /**
  * Internal dependencies
@@ -11,12 +11,14 @@ import { Button, Dialog } from '@automattic/components';
 import './style.scss';
 
 interface Props {
+	paymentMethodSummary: TranslateResult;
 	isVisible: boolean;
 	onClose: () => void;
 	onConfirm: () => void;
 }
 
 const PaymentMethodDeleteDialog: FunctionComponent< Props > = ( {
+	paymentMethodSummary,
 	isVisible,
 	onClose,
 	onConfirm,
@@ -29,16 +31,21 @@ const PaymentMethodDeleteDialog: FunctionComponent< Props > = ( {
 			onClose={ onClose }
 		>
 			<h2 className="payment-method-delete-dialog__header">
-				{ translate( 'Remove payment method?' ) }
+				{ translate( 'Remove payment method' ) }
 			</h2>
 			<p>
 				{ translate(
-					'The payment method will be removed from your account and from all the associated subscriptions.'
+					'The payment method {{paymentMethodSummary/}} will be removed from your account and from all the associated subscriptions.',
+					{
+						components: {
+							paymentMethodSummary: <strong>{ paymentMethodSummary }</strong>,
+						},
+					}
 				) }
 			</p>
 			<Button onClick={ onClose }>{ translate( 'Cancel' ) }</Button>
 			<Button onClick={ onConfirm } primary>
-				{ translate( 'Confirm' ) }
+				{ translate( 'Delete' ) }
 			</Button>
 		</Dialog>
 	);
