+++ client/me/purchases/credit-cards/payment-method-delete-dialog/style.scss
@@ -6,9 +6,9 @@
 		text-align: left;
 	}
 
-	// &.dialog {
-	// 	max-width: 440px;
-	// }
+	&.dialog {
+		max-width: 440px;
+	}
 
 	&__header {
 		font-size: 24px;
